chore: add initial .gitignore file
Introduce an initial gitignore file.

The .gitignore file consists of two sections, an empty project-specific section and the contents of the Node.gitignore file.

It is worth adding the contents of the Node.gitignore file to the file because it filters out files created by node modules, testing tools, and linters that may be used during project development from version control.

Signed-off-by: Mukli Krisztián <krisztianmukli@mukli.hu>